cdr/requirements.txt
faker
paramiko
confluent-kafka
kafka-python
